+++ src/meltano/api/controllers/dashboards_helper.py
@@ -17,7 +17,8 @@ class DashboardsHelper:
     VERSION = "1.0.0"
 
     def get_dashboards(self):
-        path = Path(Project.meltano_model_path())
+        project = Project.find()
+        path = project.root_dir("model")
         dashboardsParser = M5oCollectionParser(path, M5oCollectionParserTypes.Dashboard)
         return dashboardsParser.contents()
 
@@ -48,9 +49,10 @@ class DashboardsHelper:
         return target_dashboard[0]
 
     def save_dashboard(self, data):
+        project = Project.find()
         slug = slugify(data["name"])
         file_name = f"{slug}.dashboard.m5o"
-        file_path = Path(Project.meltano_model_path()).joinpath(file_name)
+        file_path = project.root_dir("model", file_name)
         data = MeltanoAnalysisFileParser.fill_base_m5o_dict(file_path, slug, data)
         data["version"] = DashboardsHelper.VERSION
         data["description"] = data["description"] or ""
@@ -60,21 +62,23 @@ class DashboardsHelper:
         return data
 
     def add_report_to_dashboard(self, data):
+        project = Project.find()
         dashboard = self.get_dashboard(data["dashboardId"])
         if data["reportId"] not in dashboard["reportIds"]:
             dashboard["reportIds"].append(data["reportId"])
             file_name = f"{dashboard['slug']}.dashboard.m5o"
-            file_path = Path(Project.meltano_model_path()).joinpath(file_name)
+            file_path = project.root_dir("model", file_name)
             with open(file_path, "w") as f:
                 json.dump(dashboard, f)
         return dashboard
 
     def remove_report_from_dashboard(self, data):
+        project = Project.find()
         dashboard = self.get_dashboard(data["dashboardId"])
         if data["reportId"] in dashboard["reportIds"]:
             dashboard["reportIds"].remove(data["reportId"])
             file_name = f"{dashboard['slug']}.dashboard.m5o"
-            file_path = Path(Project.meltano_model_path()).joinpath(file_name)
+            file_path = project.root_dir("model", file_name)
             with open(file_path, "w") as f:
                 json.dump(dashboard, f)
         return dashboard